fix: update the passed-in values for AddressDaoTest and MenuDatabaseTests to reflect the data generator refactors

+++ app/src/androidTest/java/com/littlelemon/application/address/data/local/dao/AddressDaoTests.kt
@@ -102,7 +102,7 @@ class AddressDaoTests {
         val count = dao.getAddressCount()
 
         // Assert
-        assertEquals(0L, count)
+        assertEquals(0, count)
     }
 
     @Test
@@ -116,6 +116,6 @@ class AddressDaoTests {
         val count = dao.getAddressCount()
 
         // Assert
-        assertEquals(numAddresses.toLong(), count)
+        assertEquals(numAddresses, count)
     }
 }
